Split quickstart smoke into dedicated command
